Handle Shift JIS titles in libnetmd's netmd_request_title
This updates libnetmd to handle UTF-8 conversion of track titles in a similar way to libhimd. I've made the assumption here that all NetMD title data is JIS encoded, which seems to be true enough to not be an issue on my discs. Clearly, though, it's possible I'm wrong (do European discs have extended character set support?), in which case a more careful method would need to be applied.